feat: Implement LOAD_NAME and BINARY_OP(part).
LOAD_NAME is opposite to STORE_NAME, it gets the variable name from
frame->locals and increase sp.

BINARY_OP contains add, mul, sub and so on. The arguments are top two of
the stack. The sp will sub two for argument and add one for return value.
Currently only add in LongObject is complete, but since the binary_op's
framework is complete, add other methods won't be difficult.