feat: Add LOAD_FAST and STORE_FAST for function.
In function, a variable is stored on stack. So compiler can convert
variable name to the index on the stack, which will save a lot time.

LOAD_FAST almost do the same thing as LOAD_NAME, except LOAD_FAST's arg
is the index of the stack, while LOAD_NAME's arg is used to get the
name(string) and find in the dict.

Obviously, function should support arguments. The CALL instruction will
put them on the stack of the new frame, so after entering the new frame,
the arguments are placed correctly.