feat: Check if an Object is root object in gc.
Traverse the gc_list and decrease the shadow_ref. We use
TypeObject::traverse to traverse all the objects in a container. This
recursion only performs one level.

If an Object's shadow_ref is not decreased to 0, then it is must
referenced by stack, so we should not free them. Note that we can't
ensure other Objects since they may just be referenced by the root
objects.