feat: Free unreachable objects.
Use TypeObject::clear to break reference cycles. clear() do the same as
destroy, but it should ensure that after the instance starts to decrement
other references in the clear function, the destroy function of the
instance can be called safely, and the memory allocated by gc_malloc()
will not be released (still on the unreachable linked list).

GC module seems to work since gc_cnt is 0 in the end.